Improved handling of the refresh sequence.
Specifically, it now no longer shows "Loading Data" and act strange when the internet connection is very slow, and it doesn't even run that check in the case of some partial-connection scenarios where it is still attempting to determine connection status while reloading (due to an install, etc).

Also, removed the "Reloading Data" when cancelling the refresh bar (and fixed another bug cancelling the bar, as well).

Finally, made a small formatting change to the changes badge to remove redundant code.